# backend/docs/swagger.yaml
definitions:
  apiv1.MetaCallbackRequest:
    properties:
      code:
        type: string
      redirect_uri:
        type: string
      state:
        type: string
    required:
    - code
    - redirect_uri
    type: object
  modelsv1.CreateInstagramChannelRequest:
    properties:
      channel_name:
        type: string
      code:
        type: string
      owner_user_id:
        type: string
      team_id:
        type: string
    required:
    - channel_name
    - code
    - owner_user_id
    type: object
  modelsv1.CreateInstagramChannelResponse:
    properties:
      channel_id:
        type: integer
      channel_name:
        type: string
    type: object
  modelsv1.FirebaseAuthParams:
    properties:
      firebase_id:
        type: string
      firebase_token:
        type: string
      platform:
        enum:
        - email
        - google
        - facebook
        - microsoft
        type: string
    required:
    - firebase_id
    - firebase_token
    - platform
    type: object
  modelsv1.RefreshTokenParams:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  modelsv1.Team:
    properties:
      avatar_url:
        type: string
      description:
        type: string
      name:
        type: string
      owner_id:
        type: string
      status:
        type: string
      visibility:
        type: string
    type: object
  modelsv1.TeamMember:
    properties:
      id:
        type: string
      invited_by:
        type: string
      is_primary:
        type: boolean
      permissions:
        additionalProperties: true
        type: object
      role:
        type: string
      status:
        type: string
      team_id:
        type: string
      user_id:
        type: string
    type: object
  modelsv1.UpdateUserParams:
    properties:
      email:
        type: string
      id:
        type: string
    type: object
  modelsv1.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      firebase_id:
        type: string
      id:
        type: string
      memberships:
        items:
          $ref: '#/definitions/modelsv1.TeamMember'
        type: array
      name:
        type: string
      platforms:
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /auth/authenticate:
    post:
      consumes:
      - application/json
      description: Authenticate user using Firebase token and local ID, create user
        if not exists
      parameters:
      - description: Firebase Authentication Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/modelsv1.FirebaseAuthParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Authenticate user with Firebase
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logs out the user by revoking their refresh token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Logout User
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refreshes the access token using a valid refresh token
      parameters:
      - description: Refresh Token Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/modelsv1.RefreshTokenParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Refresh Access Token
      tags:
      - Authentication
  /channels/instagram/create:
    post:
      consumes:
      - application/json
      description: Creates an Instagram channel by exchanging authorization code for
        token
      parameters:
      - description: Create Instagram Channel Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/modelsv1.CreateInstagramChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/modelsv1.CreateInstagramChannelResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create Instagram Channel
      tags:
      - Channels
  /meta/callback:
    post:
      consumes:
      - application/json
      description: Exchanges authorization code for access token and fetches pages
      parameters:
      - description: Meta Callback Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apiv1.MetaCallbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Handle Meta Callback
      tags:
      - Meta
  /teams:
    post:
      consumes:
      - application/json
      description: Create a team with name and owner ID
      parameters:
      - description: Team details
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/modelsv1.Team'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/modelsv1.Team'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create team
      tags:
      - Teams
  /teams/{team_id}:
    get:
      consumes:
      - application/json
      description: Get team details by ID
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/modelsv1.Team'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get team by ID
      tags:
      - Teams
    patch:
      consumes:
      - application/json
      description: Update team details by ID
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      - description: Team details
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/modelsv1.Team'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/modelsv1.Team'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update team
      tags:
      - Teams
  /teams/all:
    get:
      consumes:
      - application/json
      description: Get a list of all teams
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/modelsv1.Team'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all teams
      tags:
      - Teams
  /teams/owner/{owner_id}:
    get:
      consumes:
      - application/json
      description: Get team details by owner ID
      parameters:
      - description: Owner ID
        in: path
        name: owner_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/modelsv1.Team'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get team by owner ID
      tags:
      - Teams
  /users/{user_id}:
    get:
      consumes:
      - application/json
      description: Get user details by ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/modelsv1.User'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user by ID
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Update user details
      parameters:
      - description: Update user data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/modelsv1.UpdateUserParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/modelsv1.User'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - Users
swagger: "2.0"